+++ src/Codec/Archive/Foreign/Common.hs
@@ -1,7 +1,9 @@
 module Codec.Archive.Foreign.Common ( intToBool ) where
 
+import           Debug.Trace
 import           Foreign.C.Types
 
 -- TODO: this should check against ArchiveFatal?
 intToBool :: CInt -> Bool
-intToBool = toEnum . fromIntegral
+intToBool 0 = False
+intToBool _ = True